requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-admin==2.0.2
django-six==1.0.5
djangorestframework==3.15.2
idna==3.7
isoweek==1.3.3
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
screen==1.0.1
six==1.16.0
sqlparse==0.5.0
steamid-converter==1.0.1
TimeConvert==3.0.13
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
validators==0.28.3
xlwt==1.3.0
arrow==1.3.0
